third_party/WebKit/LayoutTests/platform/mac-retina/virtual/mojo-loading/http/tests/security/xss-DENIED-window-name-navigator-expected.txt

CONSOLE MESSAGE: line 7: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
CONSOLE MESSAGE: line 15: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
CONSOLE MESSAGE: line 7: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
 
